pyproject.toml
[tool.poetry]
name = "radiologist-api"
version = "0.1.0"
description = ""
authors = ["Orkhan Kazymov <rayskarken@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
torch = { version = "2.2.1", source="torch"}
torchvision = { version = "0.17.1", source="torch"}

grpcio-tools = "^1.63.0"
grpcio = "^1.63.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = "secondary"
